Remove filters from paging with next_page_uri
The next_page_uri includes filters, so you don't need to continue adding them
to subsequent API requests. If the next_page_uri is being used, just pass an
empty array for the filters.

Also add a test for the correct behavior in this case, based on the real API
response.